protect backend

+++ pages/api/ingredients/[id].js
@@ -1,12 +1,18 @@
 import dbConnect from "@/db/connect";
 import Ingredient from "@/db/models/Ingredient";
+import { getServerSession } from "next-auth/next";
+import { authOptions } from "../auth/[...nextauth]";
 
 export default async function handler(request, response) {
+  const session = await getServerSession(request, response, authOptions);
   await dbConnect();
   const { id } = request.query;
 
   if (request.method === "PUT") {
     try {
+      if (!session) {
+        return response.status(401).json({ status: "Not authorized" });
+      }
       const ingredientData = request.body;
       const ingredientToUpdate = await Ingredient.findByIdAndUpdate(
         id,
@@ -25,6 +31,9 @@ export default async function handler(request, response) {
     }
   }
   if (request.method === "DELETE") {
+    if (!session) {
+      return response.status(401).json({ status: "Not authorized" });
+    }
     await Ingredient.findByIdAndDelete(id);
     response
       .status(200)
